Re-enable circular dependency checks in editor
