+++ packages/ember-rdfa-editor/rollup.config.mjs
@@ -47,18 +47,15 @@ export default [
   },
   {
     preserveSymlinks: false,
-    // TODO: stream-browserify is a mess of circular deps, so I can't let the
-    // build fail here anymore. Need to find a better way
-    //
-    // onwarn: (message, defaultHandler) => {
-    //   // fail build if circular dependencies are found
-    //   if (message.code === 'CIRCULAR_DEPENDENCY') {
-    //     console.error(message);
-    //     process.exit(-1);
-    //   } else {
-    //     defaultHandler(message);
-    //   }
-    // },
+    onwarn: (message, defaultHandler) => {
+      // fail build if circular dependencies are found
+      if (message.code === 'CIRCULAR_DEPENDENCY') {
+        console.error(message);
+        process.exit(-1);
+      } else {
+        defaultHandler(message);
+      }
+    },
     // This provides defaults that work well alongside `publicEntrypoints` below.
     // You can augment this if you need to.
     output: addon.output(),
