+++ packages/ember-rdfa-editor/src/plugins/rdfa-info/index.ts
@@ -1,8 +1,9 @@
-export { rdfaInfoPluginKey, rdfaInfoPlugin } from './plugin.ts';
+export { rdfaInfoPlugin } from './plugin.ts';
 export {
   getNodeByRdfaId,
   getNodesBySubject,
   getRdfaIds,
   getSubjects,
+  rdfaInfoPluginKey,
 } from './utils.ts';
 export * from './types.ts';
